fix(unity): 修复重复初始化内存泄漏和 long→int 类型一致性
- BoardView/Board3DView 添加 _viewInitialized 保护，防止 RestartGame 重复创建容器和池
- Tile3DView Z 轴缩放跟随动画 (min(scaleX, scaleY)) 避免消除时变扁盘
- 所有 View 层 long→int 类型对齐 Core 的 int ID (减少隐式转换开销)

+++ unity/Assets/Scripts/Bridge/Match3Bridge.cs
@@ -365,7 +365,7 @@ namespace Match3.Unity.Bridge
         /// Get tile ID at grid position.
         /// Returns -1 if no tile at position.
         /// </summary>
-        public long GetTileIdAt(Position pos)
+        public int GetTileIdAt(Position pos)
         {
             if (!_initialized) return -1;
